Rename export solutions controller
